Refactor ConfigFactory and optimize VBucket memory heap size.
This changeset refactors the ConfigFactory but not changing how it
works internally. In addition, VBucket objects are moved from int
to short and only allocate as much memory as needed.

Reviewed-on: http://review.couchbase.org/29850
Tested-by: Michael Nitschinger <michael.nitschinger@couchbase.com>

+++ src/main/java/com/couchbase/client/vbucket/config/DefaultConfig.java
@@ -143,7 +143,7 @@ public class DefaultConfig extends SpyObject implements Config {
     int rv = mappedServer;
     if (mappedServer == wrongServer) {
       rv = (rv + 1) % this.serversCount;
-      this.vbuckets.get(vbucket).setMaster(rv);
+      this.vbuckets.get(vbucket).setMaster((short) rv);
     }
     return rv;
   }
